chore: clean up code and move accessor abstraction to CrossCutting project
