+++ ogybot.Bot/Commands/Groups/Waitlist/BaseWaitlistCommand.cs
@@ -1,7 +1,7 @@
 ﻿using Discord.Interactions;
 using ogybot.Bot.Commands.Core.Base;
 using ogybot.Bot.Handlers;
-using ogybot.Domain.Accessors;
+using ogybot.CrossCutting.Accessors.Abstractions;
 using ogybot.Domain.Infrastructure.Clients;
 
 namespace ogybot.Bot.Commands.Groups.Waitlist;
