+++ ogybot.Bot/Commands/Groups/Waitlist/Implementation/WaitlistCommandsConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using ogybot.Bot.Commands.Core.Validators;
 using ogybot.Bot.Handlers;
-using ogybot.Domain.Accessors;
+using ogybot.CrossCutting.Accessors.Abstractions;
 using ogybot.Domain.Infrastructure.Clients;
 
 namespace ogybot.Bot.Commands.Groups.Waitlist.Implementation;
