+++ ogybot.Bot/Commands/Misc/PingChiefsCommand.cs
@@ -3,8 +3,7 @@ using Discord.Interactions;
 using ogybot.Bot.Commands.Core.Base;
 using ogybot.Bot.Handlers;
 using ogybot.Communication.Constants;
-using ogybot.Domain.Accessors;
-using ogybot.Domain.Infrastructure.Clients;
+using ogybot.CrossCutting.Accessors.Abstractions;
 
 namespace ogybot.Bot.Commands.Misc;
