+++ ogybot.Communication/Exceptions/FetchingException.cs
@@ -1,6 +1,4 @@
-﻿using ogybot.Communication.Constants;
-
-namespace ogybot.Communication.Exceptions;
+﻿namespace ogybot.Communication.Exceptions;
 
 public class FetchingException(string message) : OgybotException(message)
 {
