+++ ogybot.Infrastructure/ogybot.Infrastructure.csproj
@@ -10,7 +10,7 @@
     <ItemGroup>
         <PackageReference Include="Discord.Net.Core" Version="3.16.0"/>
         <PackageReference Include="Discord.Net.WebSocket" Version="3.16.0"/>
-        <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="8.0.1" />
+        <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="8.0.1"/>
         <PackageReference Include="SocketIOClient" Version="3.1.2"/>
     </ItemGroup>
